requirements.txt
python-dotenv==1.1.0
python-telegram-bot==13.13
requests==2.32.4